feat: add bottomnavbar
