+++ app/src/main/java/com/nutriast/helper/ViewModelFactory.kt
@@ -4,7 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.nutriast.data.local.UserPreference
 import com.nutriast.ui.login.LoginViewModel
-import com.nutriast.ui.profile.ProfileViewModel
+import com.nutriast.ui.main.profile.ProfileViewModel
+//import com.nutriast.ui.profile.ProfileViewModel
 import com.nutriast.ui.register.RegisterViewModel
 import com.nutriast.ui.splashscreen.SplashScreenViewModel
