+++ app/src/main/java/com/nutriast/ui/main/MainActivity.kt
@@ -4,8 +4,13 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.nutriast.R
 import com.nutriast.databinding.ActivityMainBinding
-import com.nutriast.ui.profile.ProfileActivity
+//import com.nutriast.ui.profile.ProfileActivity
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
@@ -19,13 +24,14 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         getLoggedInUser()
+        setupBottomNavigationBar()
 
-        binding.profileButton.setOnClickListener {
-            val i = Intent(this, ProfileActivity::class.java)
-            i.putExtra(ProfileActivity.EXTRA_AUTH_TOKEN, authToken)
-            i.putExtra(ProfileActivity.EXTRA_USER_ID, userId)
-            startActivity(i)
-        }
+//        binding.profileButton.setOnClickListener {
+//            val i = Intent(this, ProfileActivity::class.java)
+//            i.putExtra(ProfileActivity.EXTRA_AUTH_TOKEN, authToken)
+//            i.putExtra(ProfileActivity.EXTRA_USER_ID, userId)
+//            startActivity(i)
+//        }
     }
 
     private fun getLoggedInUser() {
@@ -33,6 +39,18 @@ class MainActivity : AppCompatActivity() {
         userId = intent.getStringExtra(EXTRA_USER_ID).toString()
         val loggedInUser = listOf(authToken, userId)
         Log.d(TAG, "getLoggedInUser: $loggedInUser")
+    }
+
+    private fun setupBottomNavigationBar() {
+        val bottomNav = binding.bottomNavView
+        val navController = findNavController(R.id.nav_host_fragment_activity_main)
+        val appBarConfiguration = AppBarConfiguration(
+            setOf(
+                R.id.navigation_home, R.id.navigation_intake_history, R.id.navigation_profile
+            )
+        )
+        setupActionBarWithNavController(navController, appBarConfiguration)
+        bottomNav.setupWithNavController(navController)
     }
 
     companion object {
