+++ app/src/main/java/com/nutriast/ui/main/home/HomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.nutriast.ui.main.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
