+++ app/src/main/java/com/nutriast/ui/main/intakehistory/IntakeHistoryViewModel.kt
@@ -1,0 +1,7 @@
+package com.nutriast.ui.main.intakehistory
+
+import androidx.lifecycle.ViewModel
+
+class IntakeHistoryViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
