+++ app/src/main/java/com/nutriast/ui/main/profile/ProfileFragment.kt
@@ -1,0 +1,46 @@
+package com.nutriast.ui.main.profile
+
+import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.nutriast.R
+import com.nutriast.data.local.UserPreference
+import com.nutriast.helper.ViewModelFactory
+
+class ProfileFragment : Fragment() {
+    private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
+
+    companion object {
+        fun newInstance() = ProfileFragment()
+    }
+
+    private lateinit var profileViewModel: ProfileViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_profile, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        profileViewModel = ViewModelProvider(this)[ProfileViewModel::class.java]
+        // TODO: Use the ViewModel
+    }
+
+    private fun setupViewModel() {
+        profileViewModel = ViewModelProvider(
+            this,
+            ViewModelFactory(UserPreference.getInstance(requireContext().dataStore))
+        )[ProfileViewModel::class.java]
+    }
+
+}
