+++ app/src/main/res/drawable/ic_baseline_food_bank_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#FFFFFF"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M12,3L4,9v12h16V9L12,3zM12.5,12.5c0,0.83 -0.67,1.5 -1.5,1.5v4h-1v-4c-0.83,0 -1.5,-0.67 -1.5,-1.5v-3h1v3H10v-3h1v3h0.5v-3h1V12.5zM15,18h-1v-3.5h-1v-3c0,-1.1 0.9,-2 2,-2V18z"/>
+</vector>
