+++ app/src/main/res/layout/activity_main.xml
@@ -6,49 +6,52 @@
     android:layout_height="match_parent"
     tools:context=".ui.main.MainActivity">
 
-<!--    <Button-->
-<!--        android:id="@+id/loginButton"-->
-<!--        android:layout_width="0dp"-->
-<!--        android:layout_height="64dp"-->
-<!--        android:layout_marginStart="32dp"-->
-<!--        android:layout_marginTop="64dp"-->
-<!--        android:layout_marginEnd="32dp"-->
-<!--        android:backgroundTint="@color/dark_gray"-->
-<!--        android:text="@string/login"-->
-<!--        android:textStyle="bold"-->
-<!--        android:fontFamily="@font/urbanist_variablefont_wght"-->
+<!--    <TextView-->
+<!--        android:layout_width="wrap_content"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:text="Main Activity"-->
+<!--        app:layout_constraintBottom_toBottomOf="parent"-->
 <!--        app:layout_constraintEnd_toEndOf="parent"-->
 <!--        app:layout_constraintStart_toStartOf="parent"-->
 <!--        app:layout_constraintTop_toTopOf="parent" />-->
 
 <!--    <Button-->
-<!--        android:id="@+id/registerButton"-->
+<!--        android:id="@+id/profileButton"-->
 <!--        android:layout_width="0dp"-->
 <!--        android:layout_height="64dp"-->
 <!--        android:layout_marginStart="32dp"-->
 <!--        android:layout_marginTop="32dp"-->
 <!--        android:layout_marginEnd="32dp"-->
 <!--        android:backgroundTint="@color/dark_gray"-->
-<!--        android:text="@string/register"-->
+<!--        android:text="@string/profile"-->
 <!--        android:textStyle="bold"-->
 <!--        android:fontFamily="@font/urbanist_variablefont_wght"-->
 <!--        app:layout_constraintEnd_toEndOf="parent"-->
 <!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toBottomOf="@+id/loginButton" />-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
 
-    <Button
-        android:id="@+id/profileButton"
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottomNavView"
         android:layout_width="0dp"
-        android:layout_height="64dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="32dp"
-        android:backgroundTint="@color/dark_gray"
-        android:text="@string/profile"
-        android:textStyle="bold"
-        android:fontFamily="@font/urbanist_variablefont_wght"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_height="wrap_content"
+        android:layout_marginStart="0dp"
+        android:layout_marginEnd="0dp"
+        android:background="?android:attr/windowBackground"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:menu="@menu/bottom_nav_menu" />
+
+    <fragment
+        android:id="@+id/nav_host_fragment_activity_main"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+        app:layout_constraintBottom_toTopOf="@id/bottomNavView"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:navGraph="@navigation/mobile_navigation" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
