+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/navigation_home"
+        android:icon="@drawable/ic_baseline_home_24"
+        android:title="@string/home" />
+    <item
+        android:id="@+id/navigation_intake_history"
+        android:icon="@drawable/ic_baseline_food_bank_24"
+        android:title="Intake History" />
+    <item
+        android:id="@+id/navigation_profile"
+        android:icon="@drawable/ic_baseline_person_pin_24"
+        android:title="@string/profile" />
+</menu>
