+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@+id/navigation_home">
+
+    <fragment
+        android:id="@+id/navigation_home"
+        android:name="com.nutriast.ui.main.home.HomeFragment"
+        android:label="@string/home"
+        tools:layout="@layout/fragment_home" />
+
+    <fragment
+        android:id="@+id/navigation_intake_history"
+        android:name="com.nutriast.ui.main.intakehistory.IntakeHistoryFragment"
+        android:label="@string/intake_history"
+        tools:layout="@layout/fragment_intake_history" />
+
+    <fragment
+        android:id="@+id/navigation_profile"
+        android:name="com.nutriast.ui.main.profile.ProfileFragment"
+        android:label="@string/profile"
+        tools:layout="@layout/fragment_profile" />
+</navigation>
