+++ app/src/main/res/values/strings.xml
@@ -24,13 +24,17 @@
     <string name="user_photo">User Photo</string>
     <string name="personal_information">Personal Information</string>
     <string name="age">Age</string>
-<!--    <string name="user_gender" tools:ignore="TypographyEllipsis">...</string>-->
-<!--    <string name="user_age" tools:ignore="TypographyEllipsis">...</string>-->
-<!--    <string name="user_weight" tools:ignore="TypographyEllipsis">...</string>-->
-<!--    <string name="user_height" tools:ignore="TypographyEllipsis">...</string>-->
-<!--    <string name="user_name" tools:ignore="TypographyEllipsis">...</string>-->
-<!--    <string name="user_email" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_gender" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_age" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_weight" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_height" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_name" tools:ignore="TypographyEllipsis">...</string>-->
+    <!--    <string name="user_email" tools:ignore="TypographyEllipsis">...</string>-->
     <string name="triple_dots" tools:ignore="TypographyEllipsis">...</string>
     <string name="logout">Logout</string>
     <string name="nutriast_logo">Nutriast Logo</string>
+    <string name="home">Home</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="intake_history">Intake History</string>
 </resources>
